+++ src/components/ui/combobox.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Check, ChevronsUpDown } from "lucide-react";
 
 import { cn } from "@/lib/utils";
-import { Button } from "@/shadcnUi/Button";
+import { Button } from "@/components/ui/button";
 import {
   Command,
   CommandEmpty,
@@ -10,8 +10,12 @@ import {
   CommandInput,
   CommandItem,
   CommandList,
-} from "@/shadcnUi/Command";
-import { Popover, PopoverContent, PopoverTrigger } from "@/shadcnUi/Popover";
+} from "@/components/ui/command";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
 
 interface ComboboxProps {
   options: { value: string; label: string }[];
